requirements.txt
pymupdf>=1.23.8
iocextract>=1.16.1
# mitreattack-python: stix2-elevator removed (caused resolution-too-deep errors)
mitreattack-python
chromadb>=0.5.0
langchain>=0.3.0
langchain-community>=0.3.0
sentence-transformers>=2.7.0
ollama>=0.1.6
pyyaml>=6.0.1
pydantic>=2.5.3
reportlab>=4.0.7
streamlit>=1.32.0
pandas>=2.0.0
xlsxwriter
# ── LLM providers (install only what you need) ───────────────────
# Anthropic Claude API (recommended for best quality):
langchain-anthropic>=0.3.0
# OpenAI API:
langchain-openai>=0.3.0
# ── Multi-agent orchestration ─────────────────────────────────────
langgraph>=0.2.0
